Separate functions into modules for clarity
